# docs/swagger.yaml
definitions:
...
  admin.AdminUpdateUserPayload:
    properties:
      email:
        description: User Email
        example: someone@example.com
        type: string
      name:
        description: User name
        example: someone
        type: string
      role:
        description: User Role
        example: admin
        type: string
    required:
    - email
    - name
    - role
    type: object
...
  course.AddCourseRequestPayload:
    description: Information that should be available when you add a course
    properties:
      abbreviation:
        description: Course Name Abbreviation
        type: string
      addCourseToken:
        description: Web Token that was appended to the link
        type: string
      description:
        description: Course Description (Can be left empty)
        type: string
      email:
        description: Contributor Email
        example: someone@example.com
        type: string
      id:
        description: Course ID
        type: string
      majabbr:
        description: Course Major Abbreviation
        type: string
      major_id:
        description: Major Id, will be set by the server
        type: string
      name:
        description: Course Name
        type: string
    required:
    - abbreviation
    - email
    - id
    - name
    type: object
  course.UpdateCourseRequestPayload:
    description: Information that should be available when you add a course
    properties:
      abbreviation:
        description: Course Name Abbreviation
        type: string
      description:
        description: Course Description (Can be left empty)
        type: string
      email:
        description: Contributor Email
        example: someone@example.com
        type: string
      id:
        description: Course ID, Provided by query
        type: string
      lecturer:
        description: Course Lecturer
        type: string
      majabbr:
        description: Course Major Abbreviation
        type: string
      major_id:
        description: Major Id, will be set by the server
        type: string
      name:
        description: Course Name
        type: string
      updateCourseToken:
        description: Web Token that was appended to the link
        type: string
    required:
    - abbreviation
    - email
    - lecturer
    - name
    type: object
...
  major.AddMajorRequestPayload:
    description: Information that should be available when you add a major
    properties:
      abbreviation:
        description: Major Name Abbreviation
        type: string
      facabbr:
        description: Major Faculty Abbreviation
        type: string
      faculty_id:
        description: Faculty Id, will be set by the server
        type: string
      major_token:
        description: Web Token that was appended to the link
        type: string
      name:
        description: Major Name
        type: string
    required:
    - abbreviation
    - name
    type: object
...
  verification.VerificationSendRequestPayload:
    description: Information that should be passed when request verify
    properties:
      email:
        description: User Email
        example: someone@example.com
        type: string
    required:
    - email
    type: object
  web.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        enum:
        - success
        - failed
        type: string
    type: object
info:
  contact: {}
  description: This is Open Couseware backend
  title: Open Courseware Application
  version: 1.0.1
paths:
...
  /admin/user/{email}:
    delete:
      consumes:
      - application/json
      description: Delete a user from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Delete User By Id
      tags:
      - admin
    get:
      description: Get a user from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Get User By Email
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Update a user from database
      parameters:
      - description: Admin Update User Payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin.AdminUpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Update User By Id
      tags:
      - admin
...
  /auth/verify/resend:
    post:
      consumes:
      - application/json
      description: Send Email Verification to user
      parameters:
      - description: Register Payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/verification.VerificationSendRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Send Email Verification
      tags:
      - auth
  /course:
    get:
      description: Retrieve a list of all available courses.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Get all courses
      tags:
      - course
    put:
      consumes:
      - application/json
      description: Add a new course with the given details
      parameters:
      - description: AddCourseToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add Course payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/course.AddCourseRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Add a new course
      tags:
      - course
  /course/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a course with the specified ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: DeleteCourseToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Delete course by ID
      tags:
      - course
    get:
      description: Retrieve a course by ID
      parameters:
      - description: ID of the course to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Get a course by ID
      tags:
      - course
    patch:
      consumes:
      - application/json
      description: Update an existing course.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateCourseToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update Course Payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/course.UpdateCourseRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Update Course
      tags:
      - course
...
  /course/major:
    get:
      description: Get a list of all majors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Get all majors
      tags:
      - course
    put:
      consumes:
      - application/json
      description: Add a new major to a faculty
      parameters:
      - description: AddMajorToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add Major payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/major.AddMajorRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.BaseResponse'
      summary: Add Major
      tags:
      - course
...
swagger: "2.0"